Family Composition

Responses to a brief questionnaire about family composition.

family_composition

Format

A data frame with 19169 rows and 11 variables:

user_id

Each participant's unique ID

sex

The participant's sex

age

The participant's age in years

momage

How old was your female parent when you were born?

dadage

How old was your male parent when you were born?

oldbro

How many older brothers do you have?

oldsis

How many older sisters do you have?

youngbro

How many younger brothers do you have?

youngsis

How many younger sisters do you have?

twinbro

How many same-age (e.g. twin) brothers do you have?

twinsis

How many same-age (e.g. twin) sisters do you have?

Source

https://psyteachr.github.io/psyteachr/data/family_composition.csv