First Impressions of Faces (Aggregated)

Aggregated data from Psychological Science Accelerator project: To Which World Regions Does the Valence-Dominance Model of Social Perception Apply? https://psyarxiv.com/n26dy. Mean ratings on 13 traits for each of 120 faces shown in 10 world regions. Face characteristics at psa001_cfd_faces.

psa001_agg

Format

A data frame with 1320 rows and 15 variables:

region

world region (Africa; Asia; Australia & New Zealand; Central America & Mexico; Eastern Europe; Middle East; Scandanavia; South America; UK; USA & Canada; Western Europe)

stim_id

target image - first character designates race, second character designates gender, followed by a unique identifier that matches Target in psa001_cfd_faces

aggressive

average rated aggression (1 to 9)

attractive

average rated attractiveness (1 to 9)

caring

average rated caringness (1 to 9)

confident

average rated confidence (1 to 9)

dominant

average rated dominance (1 to 9)

emostable

average rated emotional stability (1 to 9)

intelligent

average rated intelligence (1 to 9)

mean

average rated meanness (1 to 9)

responsible

average rated responsibility (1 to 9)

sociable

average rated sociability (1 to 9)

trustworthy

average rated trustworthiness (1 to 9)

unhappy

average rated unhappiness (1 to 9)

weird

average rated weirdness (1 to 9)

Source

https://osf.io/jkt29/

Details

Full data and analysis scripts at https://osf.io/jfwtr/