Face Characteristics

Face stimulus characteristics from Psychological Science Accelerator project: To Which World Regions Does the Valence-Dominance Model of Social Perception Apply? https://psyarxiv.com/n26dy. To be used with psa001_agg.

psa001_cfd_faces

Format

A data frame with 120 rows and 4 variables:

Target

target image - first character designates race, second character designates gender, followed by a unique identifier that matches Target in the CFD dataset

Race

target self-identified race (A = asian; B = black; L = latinx; W = white)

Gender

target sex (F = female; M = male)

Age

Estimate the approximate age of this person (in years) ranges 18.7 to 34.9

Source

https://osf.io/rzgh2/

Details

Faces from Ma, Correll, & Wittenbrink (2015). The Chicago Face Database: A Free Stimulus Set of Faces and Norming Data. Behavior Research Methods, 47, 1122-1135.

Full data and analysis scripts at https://osf.io/jfwtr/