Sensation Seeking Scale

Zuckerman M. (1984). Sensation seeking: a comparative approach to a human trait. Behavioral and Brain Sciences. 7: 413-471.

sensation_seeking

Format

A data frame with 27021 rows and 17 variables:

id

Each questionnaire completion's unique ID

user_id

Each participant's unique ID

date

Date of completion (YYY-mm-dd)

sss1

1:I would like a job that requires a lot of traveling. vs 0:I would prefer a job in one location.

sss2

1:I am invigorated by a brisk, cold day. vs 0:I cannot wait to get indoors on a cold day.

sss3

1:I get bored seeing the same old faces. vs 0:I like the comfortable familiarity of everyday friends.

sss4

0:I would prefer living in an ideal society in which everyone is safe, secure, and happy. vs 1:I would have preferred living in the unsettled days of our history.

sss5

1:I sometimes like to do things that are a little frightening. vs 0:A sensible person avoids activities that are dangerous.

sss6

0:I would not like to be hypnotized. vs 1:I would like to have the experience of being hypnotized.

sss7

1:The most important goal of life is to live it to the fullest and experience as much as possible. vs 0:The most important goal of life is to find peace and happiness.

sss8

1:I would like to try parachute jumping. vs 0:I would never want to try jumping out of a plane, with or without a parachute.

sss9

0:I enter cold water gradually, giving myself time to get used to it. vs 1:I like to dive or jump right in to the ocean or a cold pool.

sss10

0:When I go on vacation, I prefer the comfort of a good room and bed. vs 1:When I go on vacation, I prefer the change of camping out.

sss11

1:I prefer people who are emotionally expressive even if they are a bit unstable. vs 0:I prefer people who are calm and even-tempered.

sss12

1:A good painting should shock or jolt the senses. vs 0:A good painting should give one a feeling of peace and security.

sss13

0:People who ride motorcycles must have some kind of unconscious need to hurt themselves. vs 1:I would like to drive or ride a motorcycle.

sss14

0:In a good sexual relationship, people never get bored with each other. vs 1:It is normal to get bored after a time with the same sexual partner.

Source

https://psyteachr.github.io/psyteachr/data/sensation_seeking.csv