Systemizing Quotient

Reverse coded (Q#R) questions coded as strongly disagree = 2, slightly disagree = 1, else = 0. The other questions are coded as strongly agree = 2, slightly agree = 1, else = 0.

sq_data

Format

A data frame with 4745 rows and 31 variables:

user_id

Each participant's unique ID

sex

The participant's sex

age

The participant's age in years

id

Each questionnaire completion's unique ID

starttime

The time the questionnaire was started (yyyy-MM-dd HH:mm:ss)

endtime

The time the questionnaire was completed (yyyy-MM-dd HH:mm:ss)

Q01

If I were buying a car, I would want to obtain specific information about its engine.

Q02

If there was a problem with the electrical wiring in my home, I’d be able to fix it myself.

Q03R

I rarely read articles or web pages about new technology.

Q04R

I do not enjoy games that involve a high degree of strategy.

Q05

I am fascinated by how machines work.

Q06

In math, I am intrigued by the rules and patterns governing numbers.

Q07R

I find it difficult to understand instruction manuals for putting appliances together.

Q08

If I were buying a computer, I would want to know exact details about its hard disc drive capacity and processor speed.

Q09R

I find it difficult to read and understand maps.

Q10R

When I look at a piece of furniture, I do not notice the details of how it was constructed.

Q11R

I find it difficult to learn my way around a new city.

Q12R

I do not tend to watch science documentaries on television or read articles about science and nature.

Q13

If I were buying a stereo, I would want to know about its precise technical features.

Q14

I find it easy to grasp exactly how odds work in betting.

Q15R

I am not very meticulous when I carry out D.I.Y.

Q16

When I look at a building, I am curious about the precise way it was constructed.

Q17R

I find it difficult to understand information the bank sends me on different investment and saving systems.

Q18

When travelling by train, I often wonder exactly how the rail networks are coordinated.

Q19R

If I were buying a camera, I would not look carefully into the quality of the lens.

Q20R

When I hear the weather forecast, I am not very interested in the meteorological patterns.

Q21

When I look a mountain, I think about how precisely it was formed.

Q22

I can easily visualize how the motorways in my region link up.

Q23R

When I'm in a plane, I do not think about the aerodynamics.

Q24

I am interested in knowing the path a river takes from its source to the sea.

Q25R

I am not interested in understanding how wireless communication works.

Source

https://psyteachr.github.io/psyteachr/data/sq_data.csv

Details

Wakabayashi, A., Baron-Cohen, S., Wheelwright, S., Goldenfeld, N., Delaney, J., Fine, D., Smith, R., & Weil, L. (2006). Development of short forms of the Empathy Quotient (EQ-Short) and the Systemizing Quotient (SQ-Short). Personality and Individual Differences, 41(5), 929–940. https://doi.org/10.1016/j.paid.2006.03.017